+++ dart/test/http_client/breadcrumb_client_test.dart
@@ -9,7 +9,7 @@ import 'package:test/test.dart';
 
 import '../mocks/mock_hub.dart';
 
-final requestUri = Uri.parse('https://example.com');
+final requestUri = Uri.parse('https://example.com/path?foo=bar#baz');
 
 void main() {
   group(BreadcrumbClient, () {
@@ -30,8 +30,10 @@ void main() {
       final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['url'], 'https://example.com/path');
       expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.data?['http.query'], 'foo=bar');
+      expect(breadcrumb.data?['http.fragment'], 'baz');
       expect(breadcrumb.data?['status_code'], 200);
       expect(breadcrumb.data?['reason'], 'OK');
       expect(breadcrumb.data?['duration'], isNotNull);
@@ -51,8 +53,10 @@ void main() {
       final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['url'], 'https://example.com/path');
       expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.data?['http.query'], 'foo=bar');
+      expect(breadcrumb.data?['http.fragment'], 'baz');
       expect(breadcrumb.data?['status_code'], 404);
       expect(breadcrumb.data?['reason'], 'NOT FOUND');
       expect(breadcrumb.data?['duration'], isNotNull);
@@ -68,8 +72,10 @@ void main() {
       final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['url'], 'https://example.com/path');
       expect(breadcrumb.data?['method'], 'POST');
+      expect(breadcrumb.data?['http.query'], 'foo=bar');
+      expect(breadcrumb.data?['http.fragment'], 'baz');
       expect(breadcrumb.data?['status_code'], 200);
       expect(breadcrumb.data?['duration'], isNotNull);
     });
@@ -84,8 +90,10 @@ void main() {
       final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['url'], 'https://example.com/path');
       expect(breadcrumb.data?['method'], 'PUT');
+      expect(breadcrumb.data?['http.query'], 'foo=bar');
+      expect(breadcrumb.data?['http.fragment'], 'baz');
       expect(breadcrumb.data?['status_code'], 200);
       expect(breadcrumb.data?['duration'], isNotNull);
     });
@@ -100,8 +108,10 @@ void main() {
       final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['url'], 'https://example.com/path');
       expect(breadcrumb.data?['method'], 'DELETE');
+      expect(breadcrumb.data?['http.query'], 'foo=bar');
+      expect(breadcrumb.data?['http.fragment'], 'baz');
       expect(breadcrumb.data?['status_code'], 200);
       expect(breadcrumb.data?['duration'], isNotNull);
     });
@@ -160,8 +170,10 @@ void main() {
       final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
 
       expect(breadcrumb.type, 'http');
-      expect(breadcrumb.data?['url'], 'https://example.com');
+      expect(breadcrumb.data?['url'], 'https://example.com/path');
       expect(breadcrumb.data?['method'], 'GET');
+      expect(breadcrumb.data?['http.query'], 'foo=bar');
+      expect(breadcrumb.data?['http.fragment'], 'baz');
       expect(breadcrumb.level, SentryLevel.error);
       expect(breadcrumb.data?['duration'], isNotNull);
     });