+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### Dependencies
 
-- Bump Cocoa SDK from v8.3.1 to v8.3.2 ([#1350](https://github.com/getsentry/sentry-dart/pull/1350))
-  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#832)
-  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.3.1...8.3.2)
+- Bump Cocoa SDK from v8.3.1 to v8.3.3 ([#1350](https://github.com/getsentry/sentry-dart/pull/1350), [#1355](https://github.com/getsentry/sentry-dart/pull/1355))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#833)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.3.1...8.3.3)
 
 ## Fixes