+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 ### Features
 
 - Sanitize sensitive data from URLs (span desc, span data, crumbs, client errors) ([#1327](https://github.com/getsentry/sentry-dart/pull/1327))
+### Dependencies
+
+- Bump Cocoa SDK from v8.3.1 to v8.3.2 ([#1350](https://github.com/getsentry/sentry-dart/pull/1350))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#832)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.3.1...8.3.2)
+
+## Fixes
+
+- Sync missing properties to the Native SDKs ([#1354](https://github.com/getsentry/sentry-dart/pull/1354))
 
 ## 7.2.0